Renamed AddProject to ProjectForm
