+++ client/src/components/ProjectForm.js
@@ -1,0 +1,95 @@
+import React, { useState } from 'react';
+import { connect } from 'react-redux';
+import { addProject } from '../store/actions';
+import { useAuth0 } from '../auth-wrapper.js';
+
+import '../SASS/ProjectForm.scss';
+
+const ProjectForm = ({ addProject }) => {
+  const { user } = useAuth0();
+
+  const [project, setProject] = useState({
+    userId: user.id,
+    projectName: ''
+  });
+
+  const { userId, projectName } = project;
+
+  const handleChanges = e => {
+    setProject({
+      ...project,
+      [e.target.name]: e.target.value
+    });
+  };
+
+  const handleSubmit = e => {
+    e.preventDefault();
+    addProject(project).then(res => setProject({ projectName: '' }));
+  };
+
+  return (
+    <div className="project-form-wrapper">
+      <h2>Create a project</h2>
+      <form className="project-form-container" onSubmit={handleSubmit}>
+        <div className="inner-form-container">
+          <div className="project-form-left-column">
+            <label htmlFor="projectName">Project title</label>
+            <input
+              type="text"
+              value={projectName}
+              name="projectName"
+              id="projectName"
+              placeholder="Enter project title here"
+              onChange={handleChanges}
+            />
+            <label htmlFor="projectDescription">Project description</label>
+            <input
+              id="projectDescription"
+              type="text"
+              placeholder="Enter project description here"
+            />
+            <label htmlFor="teamMembers">Add team members</label>
+            <input
+              type="text"
+              placeholder="Enter team member usernames seperated by a comma"
+              id="teamMembers"
+            />
+          </div>
+
+          <div className="project-form-right-column">
+            <h3>Attach files</h3>
+            <button type="submit">Attach files</button>
+
+            <h3>Included files</h3>
+            <p>File names will populate here</p>
+            <label htmlFor="figmaLink">Figma link</label>
+            <input
+              type="text"
+              placeholder="Enter Figma link here"
+              id="figmaLink"
+            />
+            <label htmlFor="invisionLink">Figma link</label>
+            <input
+              type="text"
+              placeholder="Enter InVision link here"
+              id="invisionLink"
+            />
+          </div>
+        </div>
+        <div className="form-buttons-container">
+          <button type="submit">Publish</button>
+          <button type="button">Cancel</button>
+        </div>
+      </form>
+    </div>
+  );
+};
+
+const mapSTateToProps = state => {
+  return {};
+};
+
+export default connect(
+  mapSTateToProps,
+  { addProject }
+)(ProjectForm);
